+++ test/src/list/base/list_repository_mock.dart
@@ -1,4 +1,4 @@
 import 'package:flutter_bloc_patterns/src/list/base/list_repository.dart';
 import 'package:mockito/mockito.dart';
 
-class RepositoryMock<T> extends Mock implements Repository<T> {}
+class ListRepositoryMock<T> extends Mock implements ListRepository<T> {}