+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.2.0]
 
 * `Initial` state introduced along with `onReady` callback for the `ViewStateBuilder`,
+* `Repository` renamed to `ListRepository`,
 * Migrating to `bloc` version 0.16.1, and `flutter_bloc` version 0.22.0.
 
 ## [0.1.1]